Moved pattern_tools.h inside patterns.h
